feat: unimed-chronic 慢病管理服务增强（R4-R12）
- R4/R7: 预警状态流转校验 + 自动写入 action 记录
- R8: 报告生成流程增强（二维码+PDF异步生成+签章）
- R9: 异步就诊记录服务（Encounter）+ 档案共享
- R10: 健康检测增强（DB/Redis/Nacos/RocketMQ/HIS 五项）
- R11: 异步风险评估 + HIS 确诊同步
- R12: 筛查入组增强（RISK标签+疑似病种推断）
- 清理废弃 SQL 升级脚本
- 新增集成测试文件

+++ unimed-chronic/unimed-chronic-biz/src/main/java/org/dromara/chronic/domain/entity/ChAssessmentRule.java
@@ -8,6 +8,8 @@ import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import org.dromara.common.tenant.core.TenantEntity;
 
+import java.math.BigDecimal;
+
 /**
  * 风险评估规则对象 ch_assessment_rule
  *
@@ -26,7 +28,7 @@ public class ChAssessmentRule extends TenantEntity {
 
     private String dimensionName;
 
-    private Integer dimensionWeight;
+    private BigDecimal dimensionWeight;
 
     /**
      * 阈值配置 JSON
